refactor: extract popover base class
