+++ src/elements/link/common/block-link.scss
@@ -56,7 +56,7 @@
   ::slotted([slot='icon']),
   sbb-icon {
     // Fixes layout issues
-    // --sbb-icon-svg-display: block;
+    --sbb-icon-svg-display: block;
 
     @include sbb.absolute-center-y;
   }
