+++ src/elements/popover.pure.ts
@@ -1,3 +1,4 @@
 /** @entrypoint */
+export * from './popover/popover-base/popover-base.ts';
 export * from './popover/popover/popover.component.ts';
 export * from './popover/popover-close-button/popover-close-button.component.ts';
